requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
asyncer==0.0.7
attrs==25.3.0
autogen-agentchat==0.5.1
autogen-core==0.5.1
autogen-ext==0.5.1
azure-ai-inference==1.0.0b9
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
azure-search-documents==11.5.2
backoff==2.2.1
bidict==0.23.1
botbuilder-core==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
certifi==2025.1.31
cffi==1.17.1
chainlit==2.4.400
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
FLAML==2.3.4
frozenlist==1.5.0
fsspec==2025.3.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonpickle==1.4.2
jsonref==1.1.0
jupyter_client==8.6.3
jupyter_core==5.7.2
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.6.0
monotonic==1.6
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.3.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.70.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.0
opentelemetry-instrumentation-anthropic==0.39.0
opentelemetry-instrumentation-bedrock==0.39.0
opentelemetry-instrumentation-chromadb==0.39.0
opentelemetry-instrumentation-cohere==0.39.0
opentelemetry-instrumentation-crewai==0.39.0
opentelemetry-instrumentation-google-generativeai==0.39.0
opentelemetry-instrumentation-groq==0.39.0
opentelemetry-instrumentation-haystack==0.39.0
opentelemetry-instrumentation-lancedb==0.39.0
opentelemetry-instrumentation-langchain==0.39.0
opentelemetry-instrumentation-llamaindex==0.39.0
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.0
opentelemetry-instrumentation-milvus==0.39.0
opentelemetry-instrumentation-mistralai==0.39.0
opentelemetry-instrumentation-ollama==0.39.0
opentelemetry-instrumentation-openai==0.39.0
opentelemetry-instrumentation-pinecone==0.39.0
opentelemetry-instrumentation-qdrant==0.39.0
opentelemetry-instrumentation-replicate==0.39.0
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.0
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.0
opentelemetry-instrumentation-transformers==0.39.0
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.0
opentelemetry-instrumentation-watsonx==0.39.0
opentelemetry-instrumentation-weaviate==0.39.0
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyautogen==0.5.3
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.41.3
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.0.1
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traceloop-sdk==0.39.0
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uptrace==1.31.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
